move generatedocs under cli
